+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This changelog only contains the changes that are unreleased. For changes for in
 ### New Features
 - Add ability to override the runtime version per instance
 - Add in a unified modpack search to search across multiple platforms at once [#720]
+- Add category filtering when adding mods/shaders/worlds/resource packs
 
 ### Fixes
 - Remove old OmitStackTraceInFastThrow JVM arg